+++ app/lib/model/state/settings_state.dart
@@ -30,8 +30,7 @@ class SettingsState with SettingsStateMappable {
   final bool enableAnimations;
   final DeviceType? deviceType;
   final String? deviceModel;
-  final bool enableGlobalAnonymousAccess;
-  final bool enableTemporaryAnonymousAccess;
+  final bool shareViaLinkAutoAccept;
 
   const SettingsState({
     required this.showToken,
@@ -55,7 +54,6 @@ class SettingsState with SettingsStateMappable {
     required this.enableAnimations,
     required this.deviceType,
     required this.deviceModel,
-    required this.enableGlobalAnonymousAccess,
-    required this.enableTemporaryAnonymousAccess,
+    required this.shareViaLinkAutoAccept,
   });
 }