+++ app/lib/pages/tabs/settings_tab.dart
@@ -139,39 +139,31 @@ class SettingsTab extends StatelessWidget {
                     ),
                   if (_isWindows || _isLinux)
                     Visibility(
-                        visible: vm.settings.launchAtStartup || _isWindows,
-                        maintainAnimation: true,
-                        maintainState: true,
-                        child: AnimatedOpacity(
-                          opacity: vm.settings.launchAtStartup || _isWindows ? 1.0 : 0.0,
-                          duration: const Duration(milliseconds: 500),
-                          child: _BooleanEntry(
-                            label: t.settingsTab.general.launchMinimized,
-                            value: vm.settings.autoStartLaunchMinimized,
-                            onChanged: (b) async {
-                              await initDisableAutoStart(vm.settings);
-                              await ref.notifier(settingsProvider).setAutoStartLaunchMinimized(b);
-                              await initEnableAutoStartAndOpenSettings(vm.settings, _isWindows);
-                            },
-                          ),
-                        )),
-                  _BooleanEntry(
-                    label: t.settingsTab.general.anonymousAccess,
-                    value: vm.settings.enableGlobalAnonymousAccess,
-                    onChanged: (b) async {
-                      await ref.notifier(settingsProvider).setEnableGlobalAnonymousAccess(b == true);
-                    },
-                  ),
+                      visible: vm.settings.launchAtStartup || _isWindows,
+                      maintainAnimation: true,
+                      maintainState: true,
+                      child: AnimatedOpacity(
+                        opacity: vm.settings.launchAtStartup || _isWindows ? 1.0 : 0.0,
+                        duration: const Duration(milliseconds: 500),
+                        child: _BooleanEntry(
+                          label: t.settingsTab.general.launchMinimized,
+                          value: vm.settings.autoStartLaunchMinimized,
+                          onChanged: (b) async {
+                            await initDisableAutoStart(vm.settings);
+                            await ref.notifier(settingsProvider).setAutoStartLaunchMinimized(b);
+                            await initEnableAutoStartAndOpenSettings(vm.settings, _isWindows);
+                          },
+                        ),
+                      ),
+                    ),
                 ],
-
-                if (vm.advanced)
-                  _BooleanEntry(
-                    label: t.settingsTab.general.animations,
-                    value: vm.settings.enableAnimations,
-                    onChanged: (b) async {
-                      await ref.notifier(settingsProvider).setEnableAnimations(b);
-                    },
-                  ),
+                _BooleanEntry(
+                  label: t.settingsTab.general.animations,
+                  value: vm.settings.enableAnimations,
+                  onChanged: (b) async {
+                    await ref.notifier(settingsProvider).setEnableAnimations(b);
+                  },
+                ),
               ],
             ),
             _SettingsSection(
@@ -238,6 +230,19 @@ class SettingsTab extends StatelessWidget {
                 ),
               ],
             ),
+            if (vm.advanced)
+              _SettingsSection(
+                title: t.settingsTab.send.title,
+                children: [
+                  _BooleanEntry(
+                    label: t.settingsTab.send.shareViaLinkAutoAccept,
+                    value: vm.settings.shareViaLinkAutoAccept,
+                    onChanged: (b) async {
+                      await ref.notifier(settingsProvider).setShareViaLinkAutoAccept(b);
+                    },
+                  ),
+                ],
+              ),
             _SettingsSection(
               title: t.settingsTab.network.title,
               children: [