+++ app/lib/pages/web_send_page.dart
@@ -27,29 +27,22 @@ class WebSendPage extends StatefulWidget {
 class _WebSendPageState extends State<WebSendPage> with Refena {
   _ServerState _stateEnum = _ServerState.initializing;
   bool _encrypted = false;
-  bool _enableTemporaryAnonymousAccess = false;
-  bool _enableGlobalAnonymousAccess = false;
   String? _initializedError;
-
-
 
   @override
   void initState() {
     super.initState();
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      _init(encrypted: false, anonymousAccess: false);
+      _init(encrypted: false);
     });
   }
 
-  void _init({required bool encrypted, required bool anonymousAccess}) async {
+  void _init({required bool encrypted}) async {
     final settings = ref.read(settingsProvider);
-    _enableGlobalAnonymousAccess = settings.enableGlobalAnonymousAccess;
-    anonymousAccess = anonymousAccess | _enableGlobalAnonymousAccess;
     setState(() {
       _stateEnum = _ServerState.initializing;
       _encrypted = encrypted;
       _initializedError = null;
-      _enableTemporaryAnonymousAccess = anonymousAccess;
     });
     await sleepAsync(500);
     try {
@@ -57,9 +50,7 @@ class _WebSendPageState extends State<WebSendPage> with Refena {
             alias: settings.alias,
             port: settings.port,
             https: _encrypted,
-            anonymousAccess: anonymousAccess
           );
-      await ref.notifier(settingsProvider).setEnableTemporaryAnonymousAccess(anonymousAccess);
       await ref.notifier(serverProvider).initializeWebSend(widget.files);
       setState(() {
         _stateEnum = _ServerState.running;
@@ -264,37 +255,29 @@ class _WebSendPageState extends State<WebSendPage> with Refena {
                     Checkbox(
                       value: _encrypted,
                       onChanged: (value) {
-                        _init(encrypted: value == true, anonymousAccess: _enableTemporaryAnonymousAccess);
+                        _init(encrypted: value == true);
                       },
                     ),
                   ],
                 ),
-                Row(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Text(t.webSharePage.anonymousAccess, style: Theme.of(context).textTheme.titleMedium),
-                    const SizedBox(width: 10),
-                    Checkbox(
-                      value: _enableTemporaryAnonymousAccess,
-                      onChanged: (value) {
-                        if(_enableGlobalAnonymousAccess) {
-                          // global setting will override to temporary
-                          // so can't change this if global is on
-                          return;
-                        }
-                        _init(encrypted: _encrypted, anonymousAccess: value == true );
-                      },
-                    ),
-                    if(_enableGlobalAnonymousAccess)
-                      Text(t.webSharePage.anonymousAccessTips, style: Theme.of(context).textTheme.titleMedium),
-                  ],
-                ),
-
                 if (_encrypted)
                   Text(
                     t.webSharePage.encryptionHint,
                     style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Theme.of(context).colorScheme.warning),
                   ),
+                Row(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Text(t.webSharePage.autoAccept, style: Theme.of(context).textTheme.titleMedium),
+                    const SizedBox(width: 10),
+                    Checkbox(
+                      value: webSendState.autoAccept,
+                      onChanged: (value) {
+                        ref.notifier(serverProvider).setWebSendAutoAccept(value == true);
+                      },
+                    ),
+                  ],
+                ),
               ],
             );
           },