+++ app/lib/provider/settings_provider.dart
@@ -39,26 +39,8 @@ class SettingsService extends PureNotifier<SettingsState> {
         enableAnimations: _persistence.getEnableAnimations(),
         deviceType: _persistence.getDeviceType(),
         deviceModel: _persistence.getDeviceModel(),
-        enableGlobalAnonymousAccess: _persistence.getEnableGlobalAnonymousAccess(),
-        enableTemporaryAnonymousAccess: _persistence.getEnableTemporaryAnonymousAccess(),
+        shareViaLinkAutoAccept: _persistence.getShareViaLinkAutoAccept(),
       );
-
-
-  Future<void> setEnableGlobalAnonymousAccess(bool v) async {
-    await _persistence.setEnableGlobalAnonymousAccess(v);
-
-    state = state.copyWith(
-      enableGlobalAnonymousAccess: v,
-    );
-  }
-
-  Future<void> setEnableTemporaryAnonymousAccess(bool v) async {
-    await _persistence.setEnableTemporaryAnonymousAccess(v);
-
-    state = state.copyWith(
-      enableTemporaryAnonymousAccess: v,
-    );
-  }
 
   Future<void> setAlias(String alias) async {
     await _persistence.setAlias(alias);
@@ -197,6 +179,14 @@ class SettingsService extends PureNotifier<SettingsState> {
     await _persistence.setDeviceModel(deviceModel);
     state = state.copyWith(
       deviceModel: deviceModel,
+    );
+  }
+
+  Future<void> setShareViaLinkAutoAccept(bool shareViaLinkAutoAccept) async {
+    await _persistence.setShareViaLinkAutoAccept(shareViaLinkAutoAccept);
+
+    state = state.copyWith(
+      shareViaLinkAutoAccept: shareViaLinkAutoAccept,
     );
   }
 }